rename InstallationButton to InstallationActions
