+++ client/src/components/Toolshed/RepositoryDetails/InstallationActions.vue
